rewrites output to preserve spaces in Learn
ven though it's in code quotes, the double spaces are getting
condensed to single spaces. I tried fixing this by inserting a unicode
non-breakable space, but that didn't comapre equal to a space if it
was copied/pasted. So here's this ugliness, which displays properly,
and copy/pastes as expected.